rename trips and fix circular dependency issue
