+++ client/src/redux/reducers/trips/service.js
