+++ client/src/redux/store.js
@@ -2,7 +2,7 @@ import { configureStore } from '@reduxjs/toolkit'
 import mapReducer from './reducers/mapSlice'
 import modalsReducer from './reducers/modalsSlice'
 import preferencesReducer from './reducers/preferencesSlice'
-import tripReducer from './reducers/trip/tripSlice'
+import tripsReducer from './reducers/trips/tripsSlice'
 import usersReducer from './reducers/users/usersSlice'
 import viewReducer from './reducers/viewSlice'
 
@@ -12,7 +12,7 @@ export const store = configureStore({
     preferences: preferencesReducer,
     modals: modalsReducer,
     map: mapReducer,
-    trip: tripReducer,
+    trips: tripsReducer,
     users: usersReducer,
   },
 })
