+++ server/controllers/Controllers.js
@@ -3,8 +3,8 @@ const TripController = require('./TripController')
 const StageController = require('./StageController')
 
 const userController = new UserController()
-const tripController = new TripController()
 const stageController = new StageController()
+const tripController = new TripController(stageController)
 
 const controllers = {
   userController: userController,
